app/requirements.txt
asgiref==3.4.1
attrs==21.2.0
redis
celery
az-iranian-bank-gateways==1.6.12
cached-property==1.5.2
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
convert-numbers==0.4
cryptography==3.4.8
decorator==4.4.2
defusedxml==0.7.1
Django==3.2.6
django-countdowntimer-model==0.0.7
django-recaptcha==2.0.6
django-render-partial==0.4
django-rest-framework==0.1.0
django-tinymce==3.3.0
djangorestframework==3.12.4
humanize==3.11.0
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.5
isodate==0.6.0
moviepy==1.0.3
oauthlib==3.1.1
platformdirs==2.3.0
proglog==0.1.9
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.1.0
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
sweetify==2.1.0
tqdm==4.62.2
urllib3==1.26.6
zeep==4.1.0
psycopg2